Positioning on larger screen
Supported positioning experimental components on a screen larger than the Panel on which the experiment is defined.

Panel components are defined using proportional units, and these are then used to acquire pixel values for positioning elements on the larger Window.

This setup allows for a Window to be full screen and displayed on monitors of different resolutions, and the central Panel scaled as required thereon, while allowing the relative positions and sizes of the components to remain constant.